Implement review changes for couchbase.
The two most important changes made were that the reliance on environment
variables to set secrets was removed, and the pillowfight code was broken
out into an ocs_ci/ocs/pillowfight.py file to mirror how this is also
done in ripsaw.
