+++ ocs_ci/templates/workloads/couchbase/server/pillowfight/basic-pillowfight.yaml
@@ -1,0 +1,17 @@
+apiVersion: batch/v1
+kind: Job
+metadata:
+  name: pillowfight-rbd-simple
+spec:
+  template:
+    metadata:
+      name: pillowfight
+    spec:
+      containers:
+      - name: pillowfight
+        image: couchbaseutils/pillowfight:v2.9.3
+        command: ["cbc-pillowfight",
+                  "-U", "couchbase://cb-example-0000.cb-example.couchbase-operator-namespace.svc:8091/default?select_bucket=true",
+                  "-I", "10000", "-m", "1000", "-M", "1000",
+                  "-c", "10000", "-T", "-t", "1", "-u", "Administrator", "-P", "password"]
+      restartPolicy: Never
